pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "simple-agent-pipeline"
version = "0.1.0"
description = "Minimal Temporal-orchestrated debt collection pipeline."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.52.0",
  "fastapi>=0.115.12",
  "pydantic>=2.11.3",
  "python-dotenv>=1.1.0",
  "temporalio>=1.11.1",
  "tiktoken>=0.9.0",
  "uvicorn[standard]>=0.34.2",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
